requirements.txt
amqp==5.3.1
asgiref==3.11.1
autopep8==2.3.2
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.7
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.9.0
django-debug-toolbar==6.3.0
django-filter==25.2
django-templated-mail==1.1.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.95.0
flower==2.0.1
humanize==4.15.0
idna==3.11
kombu==5.6.2
mysqlclient==2.2.8
oauthlib==3.3.1
packaging==26.1
pillow==12.2.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
pycodestyle==2.14.0
pycparser==3.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2026.1.post1
redis==7.4.0
requests==2.33.1
requests-oauthlib==2.0.0
setuptools==82.0.1
six==1.17.0
social-auth-app-django==5.8.0
social-auth-core==4.8.6
sqlparse==0.5.5
tornado==6.5.5
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
wheel==0.47.0
